separate into different files, boundary conditions, almost working
