+++ headers/compact_tension.h
@@ -1,0 +1,107 @@
+namespace HMM {
+
+	template <int dim>
+	class CompactTension: public ProblemType<dim>
+	{
+		public:
+			CompactTension (boost::property_tree::ptree input)
+      {
+        input_config = input;
+      }
+
+			void make_grid(parallel::shared::Triangulation<dim> &triangulation)
+      {	
+				MeshDimensions mesh = this->read_mesh_dimensions(input_config);
+					
+				// Generate block with bottom in plane 0,0. Strain applied in z axis	
+				Point<dim> corner1 (0,0,0);
+				Point<dim> corner2 (mesh.x, mesh.y, mesh.z);
+
+				std::vector<uint32_t> reps {mesh.x_cells, mesh.y_cells, mesh.z_cells}; 
+				GridGenerator::subdivided_hyper_rectangle(triangulation, reps, corner1, corner2);
+      }
+
+			void define_boundary_conditions(DoFHandler<dim> &dof_handler)
+			{
+				typename DoFHandler<dim>::active_cell_iterator cell;
+
+				for (cell = dof_handler.begin_active(); cell != dof_handler.end(); ++cell) {
+        	double eps = cell->minimum_vertex_distance();
+					double delta = eps / 10.0;
+					for (uint32_t face = 0; face < GeometryInfo<3>::faces_per_cell; ++face){
+						for (uint32_t vert = 0; vert < GeometryInfo<3>::vertices_per_face; ++vert) {
+
+							// Point coords
+							double vertex_z = cell->face(face)->vertex(vert)(3);
+
+							// is vertex at base
+							if ( abs(vertex_z-0.0) < delta ){ 
+								for (uint32_t i=0; i<dim; i++){
+									fixed_vertices.push_back( cell->face(face)->vertex_dof_index(vert, i) );
+								}
+							}
+							
+							// is vertex on top	
+							if ( (vertex_z-mesh.z) < delta){
+								loaded_vertices.push_back( cell->face(face)->vertex_dof_index(vert, 2) );
+							}
+						}
+					}
+				}
+			}
+
+			std::map<types::global_dof_index,double> set_boundary_conditions(double t)
+			{
+				// define accelerations of boundary verticies
+				std::map<types::global_dof_index, double> boundary_values;
+				types::global_dof_index vert;
+
+				// fixed verticies have acceleration 0
+				for (uint32_t i=0; i<fixed_vertices.size(); i++){
+					vert = fixed_vertices[i];
+					boundary_values.insert( std::pair<types::global_dof_index,double> (vert, 0.0) );
+				}
+
+				// apply constant strain to top 
+				double strain_rate = input_config.get<double>("problem type.strain_rate");
+
+				double current_length = mesh.z * (1.0 + strain_rate * t);
+				double velocity = strain_rate * current_length;
+
+				for (uint32_t i=0; i<loaded_vertices.size(); i++){
+	       	vert = loaded_vertices[i];
+  	     	boundary_values.insert( std::pair<types::global_dof_index,double> (vert, velocity) );
+				}
+			
+				return boundary_values;
+			}
+
+			std::map<types::global_dof_index,double> boundary_conditions_to_zero()
+			{
+				std::map<types::global_dof_index, double> boundary_values;
+				uint32_t vert;
+
+				for (uint32_t i=0; i<fixed_vertices.size(); i++){
+					vert = fixed_vertices[i];
+					boundary_values.insert( std::pair<types::global_dof_index,double> (vert, 0.0) );
+				}
+
+				for (uint32_t i=0; i<loaded_vertices.size(); i++){
+					vert = loaded_vertices[i];
+					boundary_values.insert( std::pair<types::global_dof_index,double> (vert, 0.0) );
+				}
+
+				return boundary_values;
+			}
+
+
+
+		private:
+			boost::property_tree::ptree input_config;
+			MeshDimensions							mesh;
+
+			std::vector<uint32_t>			fixed_vertices;
+      std::vector<uint32_t>			loaded_vertices;
+	};
+
+}
